feat(ai-chat,likes,version-management,female-retention,sometime-articles): add v2 pages with BFF proxy
Extract legacy pages and create v2 variants with patchAdminAxios.
Version-management v2 removes useAuth dependency. All pages get
server component routers gated on feature flags.
